app/requirements.txt
streamlit==1.30.0
mistralai==0.0.11
python-dotenv==1.0.1
gTTS
elevenlabs
stftpitchshift==1.5.1
torchcrepe
setuptools
wheel
httpx==0.23.0
faiss-gpu
fairseq
gradio==3.34.0
ffmpeg-python
praat-parselmouth
pyworld
numpy==1.23.5
i18n
numba==0.56.4
librosa==0.9.2
mega.py
gdown
onnxruntime
pyngrok==4.1.12
torch
ffmpeg